[FIX]sap.ui.rta: Show BusyIndicator to block app during reset of changes
When the user presses "reset" in RTA the process can take a long time.
If the user closes the window or make further changes in the meantime,
this can lead to inconsistencies. This change adds a BusyIndicator to
make it clear to the user that there is a process running and block
application interaction.

Change-Id: I4bfdb65d8aa55e610b0977cbbb7cc3a4027f2129
Signed-off-by: Vitor Seifert <vitor.eduardo.seifert.bazzo@sap.com>