[FIX] sap.m.SegmentedButton: Now correctly goes into overflow when used inside OverflowToolbar
The issue was that updating properties of a SegmentedButtonItem before
an item is added to the aggregation "items" does not fire a "_change",
so the OverflowToolbar can understand that it's updated. The "_change"
event handler was registered afterwards.

There might be no any properties set to the SegmentedButtonItem when adding
it to the aggregation, so registering the event handler earlier (i.e.
SegmentedButton.prototype.init) won't solve entirely the issue. That's why
this fix fires the event upon any change of "items" aggregation separately
from property updates.

Change-Id: If997866ac062abad02389aeb14934e8a103375be
BCP: 1780513348
(cherry picked from commit 935c11e6d925cebe65c0c6d9649f0539314f2570)
CR-Id: 002075125800000023702018