[INTERNAL] Grunt-extensions with different paths for test and src directory
Up to now the grunt config expects one path with the child
directories
* src
* test

For libraries with a different structure it is currently not possible to
use grunt-extension.js

The changes allow to specify individual "src" and "test" paths for a
library. The original "path" component is normalized to "src" and
"test" in the Gruntfile.js.

Example:
const path = require('path');

module.exports = function(grunt, gruntData) {
	gruntData.allLibraries.push(
		{
			name: 'some.lib',
			src: path.resolve(__dirname, '../../some.lib/src/main/js'),
			test: path.resolve(__dirname, '../../some.lib/src/test/js')
		}
	);
};

Change-Id: Ieb44c976aa429176e65db3bd939b8ae055c1dc36