[FIX]sap.ui.rta: Show BusyIndicator to block app during reset of changes
When the user presses "reset" in RTA the process can take a long time.
If the user closes the window or make further changes in the meantime,
this can lead to inconsistencies. This change adds a BusyIndicator to
make it clear to the user that there is a process running and block
application interaction.

CR-Id: 002075125800005242582017
BCP: 1780513374

Change-Id: Ic13d79b503072b9b64f275b121c024cc1c8891bb
Signed-off-by: Vitor Seifert <vitor.eduardo.seifert.bazzo@sap.com>