[FIX]sap.ui.rta: Show BusyIndicator to block app during reset of changes
When the user presses "reset" in RTA the process can take a long time.
If the user closes the window or make further changes in the meantime,
this can lead to inconsistencies. This change adds a BusyIndicator to
make it clear to the user that there is a process running and block
application interaction.

CR-Id: 002075125800005242572017
BCP: 1780513374

Change-Id: Ic7e567f58ced95bb6e56e55892ce1c0a4e05575f