[FIX] Update theme URLs with query parameters on RTL change
When a library theme URL contained a query parameter, e.g. when using the
"versionedLibCss" configuration and the RTL mode has been changed, the
corresponding library.css / library-RTL.css files were not changed
properly.
This has been fixed by matching the filename and query separately so
that the check for a "library" or "library-RTL" file works again.
When building the new URL, the query is again taken into account.

Change-Id: Ie871fac6263ce9cb423728cbfe135f490e32aff7
BCP: 1880000086