[FEATURE] sap.m.ColorPalette: new control implemented.
The color palette represents a range of colors for easy selection. The
colors are predefined and they can be configured, based on the use case.

The palette can be rendered in an embedded mode (simple form) by use of
the sap.m.ColorPalette or as a popover/dialog by use of the
sap.m.ColorPalettePopover.

JIRA: BGSOFUIBALKAN-3907

Change-Id: I094d34452dc6f25e358b02a7f94197424c6b42d8