[FEATURE] Core: added config option to specify init module
The bootstrap tag now allows to specify the init module via the
attribute data-sap-ui-oninit. When prefixing the module name with
"module:" the bootstrap identifies to require the given module.

The init module will be loaded via a timeout to ensure that the boot
step is completed before the module is required and loaded. This will
also ensure an async behavior for the init module.

Example usage:

<script id='sap-ui-bootstrap'
        data-sap-ui-oninit='module:sap/ui/test/Main'
        data-sap-ui-resourceroots='{"sap.ui.test": "./}'>
</script>

In addition, the Core#_executeOnInit has been extended looking up
specified functions instead of just evaluating the code. If you just
write a String without "module:" prefix and without brackets it will try
to lookup this function and execute it, e.g.:

<script>
    window.my = {
      initfn: function() { ... }
    };
</script>

<script id='sap-ui-bootstrap'
        data-sap-ui-oninit='my.initfn'>
</script>

or directly a function reference like this:

<script>
    function initfn() { ... }
</script>

<script id='sap-ui-bootstrap'
        data-sap-ui-oninit='initfn'>
</script>

or directly as code:

<script>
    function initfn() { ... }
</script>

<script id='sap-ui-bootstrap'
        data-sap-ui-oninit='initfn();'>
</script>

Change-Id: I64eb1926cdb494328214e44f0c2b547d018523e5