[INTERNAL][FEATURE] Control sample check for unlisted sample files
All files of a sample should be listed in the sample's Component.js so that
 * The file can be browsed in the demo kit
 * The downloaded sample contains all files to run it in another environment

This check hooks into the jQuery.ajax and jQuery.sap.sjax calls to detect all files loaded by the sample.
Then it compares it against the list of files listed in the component.js to make sure that no files are missing.

Files that are referenced but not found on disk are ignored as this can potentially happen and is not an error.
Locale-specific files like "i18n_en_US.properties" are ignored as well to save some processing time.

Change-Id: I62b142daceb739eff3aecaf43bc50a363c096822