[FIX] sap.f.Avatar: DetailBox opens when avatar used with list binding
The problem:
When the Avatar control is used in a list for example in a
CustomListItem utilizing list binding and a sap.m.LightBox control is
added in the detailBox aggregation the press event does not open the
LightBox control.

The root cause:
List binding internally clones the template control and the LightBox in
the aggregation but the "press" event listener points to the instance
used by the list binding as original (which is not rendered in the dom)
thus clicking one of the rendered instances of the Avatar calls the
listener on the one used for template and this fails with a side effect
error of not having imageSource.

The solution:
Override the clone method to transfer the correct event listener to the
cloned instance of the Avatar control.

Change-Id: I85de84fe7b663d1782844daaaea1fe937e92b69d
BCP: 1880004603
FIXES: https://github.com/SAP/openui5/issues/1825