[FEATURE] ui5loader: support for 'bundles' configuration
With this change, the loader can be made aware of existing bundles and
their content. If a module is known to be contained in a bundle, the
loader will require that bundle before the module itself is required.

While the configuration allows a module to be listed in multiple
bundles, the loader will only remember the one configured last. There's
no strategy implemented to decide what bundle to load in case of
overlapping bundles.

Example config:

  sap.ui.loader.config({
    bundles: {
      'bundle1': ['module1', 'module2', ...],
      'bundle2': ['moduleX', 'moduleY', ...]
    }
  });

or - UI5 specific - with extensions:

  sap.ui.loader.config({
    bundlesUI5: {
      'bundle1.js': ['Component.js', 'manifest.json', ...],
      'bundle2.js': ['Page.view.xml', 'Page.controller.js', ...]
    }
  });

Note: Loading of bundles is currently only triggered by requests for JS
modules, not by requests for other resource types (like views or
fragments or manifests). But the bundle can contain such resource types.

Note: Bundle configuration can be combined with dependency cache
information. But as dependency cache information is currently evaluated
only after a module has been requested, results might not be satisfying
(no parallel loading of bundles).

Change-Id: I5d90d386a8d73d63aab88a6c991678fa5a4437b9