[FEATURE] sap.ui.core: support for <base> tag
Before this change, most code locations in UI5 used window.location.href
or parts thereof when resolving relative URLs. But in most cases, the
URLs that UI5 deals with are at the end resolved by the browser (e.g.
when theming creates a link tag or when the Image control refers to an
image resource). And the browser resolves any relative URL not relative
to window.location but relative to document.baseURI. 

With this change, the most relevant places where relative URLs are
resolved, have been adapted to document.baseURI:

- style sheet handling in (sap/ui/dom|jQuery.sap.)includeStyleSheet
- in the AppCacheBuster (which does a lot of URL rewriting)
- in the ui5loader (which compares URLs for equality or has to calculate 
  source(Mapping)URLs etc.)
- in the RequestRecorder
- in the WebSocket utility
- in the Manifest class used for components

To support the new approach also in IE11, a polyfill for
Node.prototype.baseURI has been added as well. It is integrated into the
bootstrap (sigh) as ui5loader.js already needs it.

Obviously, the change of the URL resolution is not compatible, but as
the <base> tag was not supported before and caused a lot of trouble when
being used, the risk for regressions in productive code is assumed to be
very low (hopefully zero).

Note: a first draft of this change was developed before ui5loader.js
was integrated into master. At that point in time, the loader in
jquery.sap.global still made use of URI.js. That's the reason why this
change also makes some stubs for URI.prototype.search more robust as
they otherwise made module loading fail.

Change-Id: I32f767a85d3d5aefc8c1eabe3ce6eea0419eba83
JIRA: CPOUIFPHOENIXCORE-1750