[FIX] Core#applyTheme: avoid flash of unstyled content (FOUC) #2
In case of inserting the link for the themes manually and having an
inconsistency between the theme in the configuration of the Core and the
actual theme of the link the applyTheme would be executed and the links
would get marked with the foucmarker which later on would be removed by
the ThemeCheck again. This is fixed by an extension of
jQuery.sap.includeStyleSheet to remove that attribute again when not
including another link.

Change-Id: I192072c60e16cfd6c4a621264d9c9f2a9ee491f1