[FIX] Core#applyTheme: avoid flash of unstyled content (FOUC) #2
In case of inserting the link for the themes manually and having an
inconsistency between the theme in the configuration of the Core and the
actual theme of the link the applyTheme would be executed and the links
would get marked with the foucmarker which later on would be removed by
the ThemeCheck again. This is fixed by an extension of
jQuery.sap.includeStyleSheet to remove that attribute again when not
including another link.

Change-Id: I195c03ba1c89bf11af07b3d0de5d6678c5a8e48d
CR-Id: 002075125800000383122018
BCP: n/a