[FIX] sap.f.DynamicPage: Expand header button tabbing order fixed
BCP: 1770574236

Problem: The epxand button is part of the title, which is before the
header, causing the focus to return on tab to the first focusable
element in the header when snapped and also expands the header.

Solution: Exclude the header from the tab chain in case the header is
snapped and resotre it back upon expanding. This is done only when there
are both title and header.

Change-Id: I09593409c25dd2eada965283b4738870b7c65365