[FIX] sap.uxap.ObjectPageLayout: Scroll position is now calculated correctly
Often times the scroll position is a float number, which
sometimes misses the subsection it should be scrolled to with
less than 1px. This is why we introduce rounding. Otherwise, if
we scroll to a given subsection (for e.g located on the 3000px),
the scrollTop could have a value of 2999.978, which meant that
the _getClosestScrolledSectionId method would return the ID of
the previous subsection and the one we scrolled to.

In addition to this, sometimes the last subsection could not be
reached for the same reason. This means that if an app depends
on the _getClosestScrolledSectionId method for updating it's URL,
(like the new Demokit), the resulting URL would be incorrect.

Change-Id: Id9229e4502ebf4661f6ed59947bde1669387b0d9