[INTERNAL] sap.ui.component: Async loading of modules derived from manifest
This change only applies When creating a component via the
sap.ui.component factory using the "manifest" property.
After all dependencies have been loaded, the manifest is analyzed to
detect modules which will be required by the component instance in order
to load them before the instance is created.
If the underlying loader supports async loading, those modules (and
their dependencies) will be loaded async. This is done to prevent
possible sync request that might happen later on.

This change includes detection for the following UI5 core specific
entries within the manifest:
- /sap.ui5/rootView/type (View class (JSON/XML/HTML/JS/Template)
- /sap.ui5/routing/config/routerClass (Router class)
- /sap.ui5/models/* (Model classes)

Change-Id: I299377633ea42d87aa58707431ace314be5c0ded