[FIX] model.ODataContextBinding: use preliminary context
Usually when creating a preliminary context the actual context gets loaded
asyncly.

There are some cases where the actual context is already available and
the callback function gets executed synchronously. In this case, a
preliminary context was still expected which is wrong.

Hence, a check whether it is preliminary or not is required.

Change-Id: Id7f0f81d52c3fcabe32be962d794d6175e7a7f89
BCP: 1880067605