[INTERNAL][FIX] sap.m.ColorPalette - explored app sample error fixed.
When a ColorPalettePopover sample is used and one of the popovers is open,
the it is not destroyed upon exit of the view.
This causes a duplicate id exception when returning back from the source
pane.

The embedded use-case is extracted into a separate explored app sample.

Small API documentation improvement.

Change-Id: Iea817dea8c74060907f5ee64f2dfe8c1bc19bce4