[INTERNAL][FIX] sap.uxap.ObjectPageSection: redundant margin removed
Issue:
when there is no AnchorBar, the first section gets margin-top: 2rem.
However, if that section does not have a title, the space between the section content
and the header becomes too much

Solution:
When the title is not visible, the margin is not added to the first section.

Change-Id: I811b6e7bb76cbd6a37b773bd07003ccc34e1dcf1
BCP: 1880050248