[FIX] uxap.BlockBase: blocks outside the ObjectPageSubSection are loaded
Issue:
When enableLazyLoading is on,
the loading of blocks outside of ObjectPageSubSection is delayed,
but they will never be loaded as the connectToModel mechanism only works
through the ObjectPageSubSection.

Exptected:
Only blocks, that are direct children of ObjectPageSubSection
should be target of the lazy loading.

Solution: checking for the block`s parent

Change-Id: I3bb006362aa659f9e617d4fad92dba750d89033c
Fixes: https://github.com/SAP/openui5/issues/1853
BCP: 1880061546