[INTERNAL][FIX] sap.uxap.ObjectPageLayout: Removed redundant invocation
The call to recalculate the header heights on content resize was introduced with change Idfc099e7ca2b9fc1cfe50d9c8b157cb48dcda34c
specifically for support of the Dynamic pinnable header.

However, the call is not needed when the classic ObjectPageHeader is used.
Moreover, the extra calculations cause noticeable slowdown upon lazy-loading of multiple sections during scroll.
Removed the call for the case where it is not needed.

Change-Id: I28c4173fd5870c0016f7d2fc91cba81c37d48af7