[INTERNAL][FIX] sap.m.Select: Unnecessary event marking removed
Marking the tab event from the Select causes the
custom tab handling of the Table to be skipped.
This marking didn't exist until several weeks ago.
It wasn't added due to a bug/BLI related to it,
but rather in addition to the new custom tab handling
methods for the Select several weeks ago. This would
mean, that the best action here is to remove it

Change-Id: I6490d8f2ebabdfe052a9b5deaac10d9a9bea638f
BCP: 1880060190