[INTERNAL] sap.ui.support: Functional minor changes to the UI
1. The "rowActionCount" is set to 2 when you
"Create a new rule" or "Duplicate" and existing one and
turns to 1 when there are no temp rules

2. The tabs of the "Issues" and "Rules" have been switched

3. Severity Icons in the "Details" on the "Issues" view
have been positioned accordingly - when there was a "warning" status
the icons was not positioned right so it was cut off.

4. The single table column in the Issues view is no longer resizable
- it was totally annoying

5. Fixed the focus width of the "settings" and "open documentation"
buttons in the top right by margin and padding.
- I've also added a Toolbar to the top right content - so now it's
responsive and will resize appropriately.

6. "Issues" or "Rules" view the buttons now have a highlight option - so
that the user knows which view and button is active - all we need is a
color for the highlight, I was thinking something along the lines of
"sapMBtnActive"

Change-Id: I49771ca1c8270f26be97179e7ea721a189974eca