[FEATURE] sap.uxap.AnchorBar: Dependencies
Removed dependencies where the Renderer has a dependency to its control.
In order to use proper modules a control needs to have a dependency to
its Renderer.
Therefore the dependency has been removed to avoid cyclic dependencies.
This dependency was an enum within the AnchorBar control which has been
moved to the AnchorBarRenderer.

Change-Id: I11b82334cdd10ec9c750b5eae14d834bae7c4639