[FEATURE] sap.uxap.ObjectPageHeader: Dependencies
Removed dependencies where the Renderer has a dependency to its control.
In order to use proper modules a control needs to have a dependency to
its Renderer.
Therefore the dependency has been removed to avoid cyclic dependencies.
This dependency was the ObjectPageHeader for instanceof checks.
The instanceof check is now done lazy.

Change-Id: Ic7acdd65e01ebb5f289e3c522d53da994a527991