[FEATURE] ComponentContainer: introduce manifest property
The manifest property can be used similar to the manifest option on the
component factory function "sap.ui.component".

It controls when and from where to load the manifest for the Component.
When set to any truthy value, the manifest will be loaded asynchronously
by default and evaluated before the Component controller, if it is set
to a falsy value other than "undefined", the manifest will be loaded
after the controller. A non-empty string value will be interpreted as
the URL location from where to load the manifest. A non-null object
value will be interpreted as manifest content.

Important: The default value of the async property of the Component
Container is false. If the manifest property has a truthy value and the
async property has not been provided in the initial settings the async
property will be set to true.

Change-Id: I2a5b21158d593bc4acd6bee9f0933fb5cfb9dbc2