[INTERNAL] sap.ui.core: Modularization of jquery.sap.* modules
With this change, the jquery.sap.* modules are separated into jQuery
independent fine granular modules, building the new foundation layer
of sap.ui.core. The new modules are marked as private for the time
being to allow further refactoring after 1.54. They're used to restore
the functionality of the old jquery.sap.* modules in a fully compatible
way.

Key ingredients:
- split jquery.sap.* modules into smaller fine granular modules
- ensure full compatibility of the old jquery.sap.* modules
- separation into a base layer (sap.base; browser & UI independent)
  and a UI layer (sap.ui)
- also keep the documentation compatible (at the cost of redundancy)
  to support tools that are based on the extracted documentation
  (SDK, WebIDE, internal refactoring tools, others)
- mark all new modules as private and add a "DO NOT USE" header

THIS IS ONLY A FIRST STEP. Take the "DO NOT USE" warnings/comments/headers
serious. References to the new modules might break in future releases.
Further refactoring is likely and planned, some of the new modules might
be merged again, others might disappear or change their exports.

Change-Id: Iea8ca7c9bab0aa7b26f8827ce5706977da161441