[FEATURE] Component: declarative API for initial components
Allows to declare the initial components in the main HTML page with a
standard HTML element. By marking the HTML element with the custom
attribute data-sap-ui-component the ComponentSupport parses the data-
attributes of the HTML element and converts them into settings for the
ComponentContainer. Example:

<div data-sap-ui-component data-name="my.component"></div>

The ComponentSupport forces async loading and manifest first. If the
async attribute is declared as false it will be overwritten and if the
attributes async and manifest are omitted they be interpret as true. In
addition, autoPrefixId is true by default and the component lifecycle is
managed by the container.

To enable the ComponentSupport the ComponentSupport module needs to be
 declared as oninit module, e.g.:

<script id="sap-ui-bootstrap" src="resources/sap-ui-core.js"
        data-sap-ui-oninit="module:sap/ui/core/ComponentSupport">
</script>

To disable the autorun of the component support module the URL parameter
sap-ui-xx-componentsupport-autorun=false can be used.

Change-Id: I34a1e28260ff07a01c0f1ed5e6598b5e5afb7369