[INTERNAL] Core: remove support for legacy library-all files
The **/library-all files used a quite complicated and outdated structure
which was very sensitive to cyclic dependencies. 

Since the introduction of the library-preload.js (instead of JSON)
files, two of the main reasons to use the library-all.js files are no
longer valid: better CSP support and better cross-origin support. The
async preload can achieve the same and is better integrated with the
runtime.

As future build tooling won't support the library-all files any longer,
support for them has been removed in the runtime with this change.

A straight-forward usage of sap-ui-core-all.js will not fail, but behave
like sap-ui-core.js, using preload=sync. Apps using that setup should
migrate to sap-ui-core.js and preload=async.

Change-Id: Id85eaa9b1652bb552b2016740263a980abd2e6a4