[FIX] Configuration#setRTL: do not avoid flash of unstyled content
Changing the RTL mode via the configuration triggers a localization
change event which updates the theme urls but it doesn't trigger the
ThemeCheck. The ThemeCheck normally checks for link tags being kept to
suppress flash of unstyled content (FOUC). As the ThemeCheck is not used
for setRTL FOUC cannot be suppressed here. Only Core#applyTheme will use
the suppress FOUC feature.

Change-Id: I579429120a584023d54969da7b9a40b2262a52c8