[INTERNAL] jquery.sap.stubs: Lazy stubs for legacy API (jquery.sap.* modules)
- Add "sap/base/util/lazyProperty.js":
Generic lazy stub for property getter. Stubs are self removing either
on a get which runs the provided callback or on a set of the according
property.

- Add "jquery.sap.stubs.js":
Apply the LazyProperties according to the generated mapping. The module
is required by jquery.sap.global.

- Remove unnecessary dependencies between jquery.sap.* modules
Those dependencies were initially kept to ensure compatibility. With the
stubs, those modules will be loaded in case someone relied on those
modules to be available without explicitly declaring the direct
dependency.

- Modify access of stubbed properties inside their own modules
In case a module is loaded as normal dependency (not via a stub), any
access to a stubbed property would trigger the stub which might cause
unwanted side-effects such as duplicate module execution.
By using "Object.getOwnPropertyDescriptor" a potential getter function
is bypassed.

Change-Id: Ie246254ea6cdf24215af7ce26503295271c58b61