[INTERNAL][FIX] CfgDefaults: unit test failed for obscure languages
When the browser reported some non-BCP47 compliant languages (e.g. 'C'), 
then the CfgDefaults.html test failed as it used a different algorithm
than Configuration.js to determine the default locale.

The logic has been aligned with Configuration.js and it has been
clarified with a note that the test does not test the logic but the
defaulting in Configuration.js in general.

A future, enhanced version of the test might mock window.navigator to
also test the fallback logic (to be tested cross-browser first).

Change-Id: I393c3701c21d448a46ad9e1fb36bd76a4d518900