[FIX] sap.ui.unified.ColorPicker: DemoKit sample improvement
The problem:
In the DemoKit sample of the control the input field for the HEX value
was too short to enter a fully valid HEX.

The root cause:
The sample was constructed using a vertical layout to wrap the
ColorPicker control and the vertical layout was used only to apply
padding for the sample. Having vertical layout with padding influenced
the color picker CSS thus making the HEX input field to not stretch to
the needed size.

The solution:
Improve the sample by removing the wrapping control and only using
the color picker control with applied margin to it.

Change-Id: I943c18aaa9736d383a6900651aa98f395d344439
BCP: 1880129256