[INTERNAL] ui5loader: Expose global require and define
The former UI5 module system exposed its AMD API only as sap.ui.define
and sap.ui.require. With the introduction of the new loader these
APIs went closer to the AMD standard. As a next step the global AMD
functions define and require are exposed as well.

Additionally a switch for the compatilibity to a simultaneously used
AMD loader is introduced. The switch can be set via URL parameter or
attribute of the bootstrap script tag. If set to true, the global
properties window.define and window.require are not touched by the
ui5loader anymore, so another loader can continue to use them.

Change-Id: I743f5239af6e30e7e8b504267e45d73c9ac9815a