[INTERNAL][FIX] sap.ui.support.TechnicalInfo: Bug with starting Support Assistant is now fixed.
* When standard location is selected and the URL from local storage is invalid a problem occured
with starting the Support Assistant. The solution that was provided was check if the URL from
storage is contained in predefined URLs and if its not the location is set to custom and the
input is filled with URL and standard URL is set to application default.

* A function _getControl was created to improve the readability

* A problem with the Opa test was fixed by setting the sizeLimit to the model.
The problem occurs on Firefox and Safari when select debug modules dialog was opened in
Opa test. The problem was caused by limited display of nodes 100 of the tree control.

Change-Id: Iaf4192abe862a2f5f6b4cf2b17fb8863631da46b