[FIX] AppCacheBuster: take <base> tag into account when resolving URLs
Before this change, AppCacheBuster used window.location.href when
resolving relative URLs. When a <base>-tag with an href attribute was
present, this lead to wrong results as the browser always uses
document.baseURI (derived from first <base href=...>), not
window.location.href.

Apps that use concrete URLs instead of hashes to represent their routing
state in the URL (a common SEO technique, e.g. using history.pushState),
need to use a <base>-tag to make their resource loading independent from
the current window.location. Due to the described shortcoming, they so
far couldn't use the AppCacheBuster to improve their caching behavior.

With this change, the AppCacheBuster uses document.baseURI instead. For
IE11 support, also a polyfill for Node.prototype.baseURI has been added.
It is integrated into the bootstrap as jquery.sap.global.js already has
to resolve relative URLs.

Obviously, the change of the URL resolution is not compatible. But as
the <base> tag was not supported before and caused a lot of trouble when
being used, the assumption is that productive code won't be affected.

Change-Id: I32f767a85d3d5aefc8c1eabe3ce6eea0419eba83
CR-Id: 002075125800001055142018
BCP: n/a