[FIX] jquery.sap.global: configure ui5loader for debug mode
With the modularization change 80b3ee3, the code that evaluates the
sap-ui-debug option and propagates it to the ui5loader, was
unintentionally removed.

This change restores it and adds some basic tests for the debug
mode.

Note: Debug mode is still not implemented when async loading via script
tags is activated.

Change-Id: I4cd4d43cc30b10dddfa45d5f46903c303655ef88
BCP: 1880151185