[FIX] core/XMLTemplateProcessor: pass correct container to ExtensionPoint
Fixes commit 68c17d282bbff2a83e367fd6b7fb0d8131ae625a

With the above mentioned commit the support for ExtensionPoints in html
nodes was fixed. However the support for Extensibility of Fragments via
CustomizingConfiguration was broken since then. With this change the
correct container is also passed if the container is a Fragment without
an own Id.

Change-Id: I98032be576d6628515b823bd903ab30c5d30e0b6
BCP: 0020751294 0000061950 2018