[FEATURE] sap.m.Table: Enhancing the stickiness of the column headers
Certain gaps with respect to focus handling and scrolling are enhanced.
When an item that is overlapped by the sticky column headers is focused,
the table takes care of scrolling appropriately so that the focused item
is fully visible in the viewport. This also applies when the inner
controls are focused in the table.
Resolves:
JIRA: CPOUIFTEAMB-1078

Change-Id: I09232234aa9ff4ef0f42bb2ef90c19906210e2e9