[INTERNAL] Table: Allow to drop on empty rows
The user is allowed to drop on empty rows, if all rows are empty and
NoData is not visible. In this case a "drop on aggregation" is
performed, like it is when dropping on NoData.
If not all rows are empty, the user is still not allowed to drop on an
empty row. But if the DropPosition is "Between" or "OnOrBetween", the
drop target is set to the last non-empty row.

Change-Id: I45953475488df3e24178ffc91ba14530bd2ee748
Resolves: CPOUIFTEAMB-1018