[INTERNAL] sap.ui.core.demokit.sample.v4.ListBinding: Avoid column header overlapping
Buttons in the table header overlapped other columns if the application
was running on small screens or with zoom factor 200%.
Use table popins for responsiveness.

http://localhost:8080/testsuite/test-resources/sap/ui/core/demokit/sample/common/index.html?component=odata.v4.ListBinding
Change window size or zoom factor up to 200% and check that there is no
overlapping of headers in the employees table.

PS1: First try
PS2: Review comments

Change-Id: I0add9e9087a843e6df905dee6574fb01f6c2cbd2
BCP: 1880179831