[INTERNAL][FIX] Demokit 2.0: Fixed DemoApps links to docu and api
Legacy links ("docs/guide/{id}" and "docs/api/symbols/{id}") did not
did not redirect properly (to "topic/{id}" and
"api/{id}/:entityType:/:entityId:"). The root cause was that views in
the app were set to load async in the manifest.json, patternMatched
event was fired before controller creation and controller would not do
the redirect. Code was moved to the Router class because it is the only
one instantiated when the patternMatched event is fired.

BCP: n/a
Change-Id: I5688045d778e438055864413ee7eff20c1b1a341
JIRA: BGSOFUIPIRIN-2545
CR-Id: 002075125900000142252018