[FIX] sap.ui.core.cache.CacheManager: IE in privacy mode is disabled.
Depending on the environment, the CacheManager determines if real
implementation should be used or a NOP one.
In the list of supported environments, there was no check for privacy/incognito
mode. Some browsers (e.g. IE) support IndexedDB, but not in a privacy mode.
Those need to be excluded from the internal list of supported environments.

More details:
https://developer.mozilla.org/en-US/docs/Web/API/IndexedDB_API#Browser_compatibility

There are is no unit test, since the property window.indexedDB cannot be
manipulated.

Change-Id: I4e15c7a8170c87f78b870fd9429dcb13e1d7ee2f
BCP: 1880137917
CR-Id: 002075125900000172882018