[INTERNAL][FIX] Core: evaluate debug mode config in ui5loader-autoconfig
As a side effect of the split of jquery.sap.global.js into ui5loader,
ui5loader-autoconfig (both to be executed unconditionally in any
bootstrap script) and other modules like jquery.sap.global (now required
during bootstrap like any other module), the evaluation of the debug
mode configuration moved to a later point in time.

This not only caused the dependencies of jquery.sap.global to be always
loaded as productive sources, but only made the debug mode fail in
custom bundles that - in debug mode - did not require any modules but
relied on the unconditionally executed code (jquery.sap.global in former
versions) to prolery handle the debug mode. 

The most prominent scenario in which debug mode now failed, is the FLP
executing only core-min-0.js in debug mode.

This change moves the evaluation of the debug mode configuration back to
a much earlier point in time (ui5loader-autoconfig.js).

Additionally, this change fixes a false positive in the DebugMode test
when it runs in an environment that has debug sources but does not use
preload files (e.g. code coverage scenario during UI5 nightly builds). 
 
Change-Id: Iba16992ace26e49a83bfba621b05aa3cdcd6d48b
BCP: 1880205771
BCP: 1880200398