[FEATURE] Component/ResourceModel: enhanceWith relative to manifest
1.) ResourceModel constructor accepts enhanceWith property
In order to specify a list of resource bundles to enhance an existing
ResourceModel, the constructor has been extended to support providing
the list of these resource bundles.

2.) Component Factory handles relativeTo for enhancing ResourceModels
For component variants the manifest is loaded from another location and
the resource model should be loaded also from the other location.
Therefore the enhanceWith property in the manifest for the resource
model accepts bundleUrlRelativeTo: "manifest" next to the bundleUrl.

3.) Enhanced the tests to ensure that the order of enhance is checked
that the latter bundle for enhancing overrides the previous bundles.

4.) Support for async/sequential processing of enhanceWith in the
constructor.

Change-Id: Id4dbe6a372e9a4cb0fe946b524d004b255ad33fc