[INTERNAL][FIX] sap.m.ObjectAttribute: Fixed ObjectAttribute link in IE
Issue:
ObjectAttribute link is not working in IE.

Root cause:
ObjectAttribute is rendered with display: inline and the parent of the
span element which is hovered has pointer-events: none. Although the
span has pointer-events: auto; this does not work in IE unless display
style is changed or position: relative is added.

Solution:
Added style position: relative so that the link can work in IE.

Change-Id: I2dd340917cd680434d88be466ee7bb42024baa95
BCP: 1880199721
CR-Id: 002075125900000206712018