[INTERNAL][FIX] sap.f.DynamicPage: Header state is now read out on focus
upon focusing the DynamicPageTitle, the state of the header is now
read out, as well as indication that the title is clickable.Before
this change, there was neither such indication, nor a way of knowing
the state of the DynamicPageTitle (expanded/collapsed) through JAWS.
This was because the aria-expanded attribute is set on the header
itself, which can not be reached via keyboard navigation.

In addition to this, this change fixes another issue as well:
- the aria-expanded attribute was updated only on scroll, and didn't
  work when clicking on the title or expand/collapse button.

Furthermore, SNAPPED_HEADER was renamed to COLLAPSED_HEADER in the
messagebundle, to keep it consistant with other controls in the
framework.

Change-Id: I185937de8975eb7ed27146e95fb9fd3ef1098fb8
BCP: 1880044119