[INTERNAL][FIX] sap.uxap.ObjectPageLayout: Ensured resumeScroll restores state
Background:
The _resumeScroll function restores the last saved scrollPosition
AND relies on the *_onScroll* listener to auto-detect the currently scrolled section and the correct snapped state (as expected by design)

Problem:
The *_onScroll* listener may NOT be called upon __resumeScroll: It will not be called if NO scrolling occurs (e.g. if the page is already in the expected scrollPosition)

Solution:
Fixed by ensuring that _onScroll is called.
(We already use the same strategy in the _initAnchorBarScroll function)

Ensured that snapped state and scroll position are restored upon _resumeScroll

Change-Id: Ia0e8df58fd37750e04ecb6f320b086980a98e4b5