[FIX] sap.m.Toolbar: Focus indication for active toolbars fixed
sap.m.Toolbar can be focused if it's "active" property is set
to "true". When it was focused however, there was no indication,
since the "outline" style was set to 0, and was only changed for
active Info Toolbars. This is incorrect and so we now apply "outline"
for all active toolbars, rather than just for active info toolbars.

Change-Id: I875cd41b2bed0705460eeda0d48722e068c52716
BCP: 1880228436