[INTERNAL] jquery.sap.script: replaced forIn implementation
Replaced jQuery.sap.forIn with each module as the
funcitonality is nearly the same.
While jQuery.sap.forIn executes the callback without a context
each uses the current list item as context for the callback.
This should not cause any effect since the callback's context is 
typically ignored as the execution focuses on the iteration.
jQuery.sap.forIn returns undefined, while each returns the array object.
But the return type of jQuery.sap.forIn calls is typically ignored.

Change-Id: I931b5fed14f25aa62fb61859f5ca83dd68f03248