[INTERNAL] sap.ui.model.odata._ODataMetaModelUtils#findIndex: return first match
The implementation of findIndex didn't return the first but the last
element that matches (code comments indicate that it should return the
first matching element). In the scenarios where it is used it doesn't
matter because there is at most one match.
Fixed it to return the first match.

PS1: fixed
PS2: removed spaces, added JIRA to commit message
PS3: review comments

Change-Id: I664001ebdd81d812d2072f9e4599578fbea2ac98
JIRA: CPOUI5UISERVICESV3-1199